backend_project/requirements.txt
Django==5.1.3
torch>=1.7
torchvision>=0.8
torchmetrics>=1.7.1
numpy>=1.18.5
opencv-python>=4.1.2
matplotlib>=3.2.2
PyYAML>=5.3.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
scipy>=1.4.1
pillow>=10.3.0
requests>=2.32.2 
thop>=0.1.1
django-cors-headers==4.0.0
psycopg2-binary==2.9.7
celery==5.3.1
redis
mlflow
minio==7.1.14
dvc